Add some clarification to What is indentation...
Please regard this PR as a suggestion...
I found this a _very_ hard page to grasp!

At first, I was expecting that after clicking on a parent in the hierarchy a second time, I would find a block with the links to children, and possibly indented links to grandchildren. But there weren't any blocks on these parent pages!

It took me a while to realize that the branch as created on the documentation page is _just one way_ to look at the relations between the separate notes.

It is totally possible to create another branch, that composes another way to look at the relations between those very same notes, e.g. in the context of a game between generations:
[[a game of Contract Bridge]]
  Team 1
    [[Parent A]]
    [[Child D]]
  Team 2
    [[Child B]]
    [[Child and Parent C]]

In the Linked References section of all those 4 'person' notes, now the alternative hierarchy can be seen.

And that's the basics of indentation! ...

However, I just don't know if adding all this extra information will clutter the documentation page too much or not.